@@ -44,7 +44,8 @@ type CacheEntry struct {
 type RequestError struct {
 	Error    error
 	Code     int
-	UserInfo interface{}
+	UserCode int
+	// TODO: add UserInfo as a kind of opaque data pointer. In that way, the user can return to the invoker any kind of additional data
 }
 
 type CacheDriver struct {
